add native `ogg_packet_replace_buffer()` function
This is needed because of how Streams in node work now, in that once an
`ogg_packet` has been read() from the DecoderStream, there's no way to know when
the consumer is done with the packet. This is a problem since libogg reuses the
"packet" space of each `ogg_packet` instance each time you call
`ogg_stream_packetout()`.

The unfortunate workaround is to pass in a new node Buffer of "bytes" size after
each packetout() call, and copy over the "packet" contents to the JavaScript
buffer before we pass it off to the user.